requirements.txt
absl-py==0.13.0
ansicon==1.89.0
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
blessed==1.18.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
flatbuffers==1.12
flyingcircus==0.1.4.0
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
jinxed==1.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
mysql-connector-python==8.0.25
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
pandas==1.2.5
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.1
python-dateutil==2.8.1
pytk==0.0.2.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools-scm==6.0.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
